server/pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.117.1",
    "google-auth>=2.40.3",
    "logging>=0.4.9.6",
    "numpy>=2.3.3",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.43",
    "tensorflow>=2.20.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.2",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "F", # Pyflakes
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E", # pycodestyle
    "W", # Warning
    # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    # https://docs.astral.sh/ruff/rules/#mccabe-c90
    "C", # Complexity (mccabe+) & comprehensions
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP", # pyupgrade
    # https://docs.astral.sh/ruff/rules/#isort-i
    "I", # isort
]
ignore = [
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E402", # module level import not at top of file
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP006", # use-pep585-annotation
    "UP007", # use-pep604-annotation
    "E741",  # Ambiguous variable name
    # "UP035", # deprecated-assertion
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401", # unused import
    "F403", # star imports
]

[tool.ruff.lint.mccabe]
max-complexity = 24

[tool.ruff.lint.pydocstyle]
convention = "numpy"
